First attempt to get data using Meet Up API
The first Meet Up Package that was tested is removed. On its place, we are currently working with meetup.api. It can be installed using pip (pip install meetup.api). It offers python 3.6 compatibility and it is easy to use. This first script calls Meet Up API to get a full list of all the event's categories that are available. Then, for a specific city, it retrieves all the events that are scheduled and sorts and writes them by category in an output file. The main goal of this script is to show how the basic calls to the Meet Up API works by using Python along with the meetup.api package.