Deleted old file 1stAPIcall.py and fixed a bug in MUAPIcall.py
The old file 1stAPIcall.py is deleted from the repository.
A bug is fixed in the MUAPIcall.py where a comma was missing in the ouput file.
Commas in the name field of the activities could affect to the output parser. Then, instead of splitting the name and the coordinate points with commas, tabulator space is used. In this way, data output will be easier to handle since the commas in the name field will not affect. Field are separated by tabulator spaces.