New version of mu_requests is released to improve MeetUp requests and to arrange activities by their category
Now activities can be obtained arranged by their category.
write_locations() and write_num_activities() are functions which create a custom csv file with activities arranged by category
A new function is included which is able to get the complete list of categories from MeetUp.
A new function is used to include automatically the MeetUp Key to the MeetUp client.
The way to perform the requests is changed to avoid and fix some JSON transcription problems.