Some improvements are introduced to meetup requesting and mapping functions
The way how data is saved into the csv files is changed. Now the name, the date and the id of the events are also saved.
Writting functions are modified in order to write properly the new custom csv files.
A new function 'write_data()' is introduced with this purpose.
A function called 'categories_parser()' is designed to parse category dictionaries in order to acces to the categories easier.
The function responsible to retrieve and save city events data is modified and renamed according to these new features.
A function called 'data_parser()' is introduced to parse JSON formatted data into a list of dictionaries which are easier to handle.
The functions that make requests to MeetUp are now more clever and detected when the key is missing.
New functions are introduced in 'mapping.py' in order to import correctly the information about the events from the brand new custom csv files.
Now, mapping functions do not need a mandatory categories list input parameter. They still can map and arrange events by category since if needed, they can retrieve categories list from MeetUp themselves.