Some bugs are fixes and the color pattern handle is improved
A new 'color_pattern_parser()' function is introduced to allow the user to easily control the color patterns of the event points plotted on the map.
Now when categories are not supplied, a local list of the MeetUp categories is called to get them offline.
A bug related with the MeetUp parameters is fixed. The list of parameters was not restored after a call and could produce errors to the next one.
Other minor changes are applied.