Wikipedia scraping script now allows to scrap districts from London and Munich
Compatibility of the Wikipedia scraping script is extended.
Information about the districts of London and Munich is included in the districts file.
'scrap_districts_population()' can iteratively call itself to perform an intelligent search for a Wikipedia page with an appropriate table with population data.
Some minor bugs are fixed.